+++ src/main/resources/mapper/TicketOrderMapper.xml
@@ -50,8 +50,6 @@
 
 
     <update id="updateTicketOrderStatus" parameterType="com.heeverse.ticket_order.domain.dto.persistence.TicketOrderUpdateMapperDto">
-    <update id="updateTicketOrderStatus"
-            parameterType="com.heeverse.ticket_order.domain.dto.persistence.TicketOrderUpdateMapperDto">
         UPDATE
             ticket_order
         SET booking_status = #{bookingStatus}
